A slightly modifief patch from Sascha Rodekamp "The positioning mechanism of the new lookups is only working with the bizznesstime theme" (https://issues.apache.org/jira/browse/OFBIZ-3430) - OFBIZ-3430
For the moment I only removed comments around an error which is useful. I guess Sascha was blocked by it during his tests and forgot to remove it. BTW I improved the error and one related (variables names explicit)

I will certainly do a review on the format later to avoid both in the same commit (if any changes are needed anyway)



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@906524 13f79535-47bb-0310-9956-ffa450edef68