A patch from Blas Rodriguez Somoza OFBIZ-3621 - XHTML validation errors (themes) XHTML validation errors
    * Unclosed INPUT
    * attributes without values (checked, selected, disabled, etc)
    * attribute values without "
    * Uppercase tags or attributes.
    * Unencoded ampersands in urls.

Other errors:

Header.ftl (tomahawk & droppingcrumbs)
Line 108. Wrong markup. Div can't be inside <a>
<a href="<@ofbizUrl>${logoLinkURL}</@ofbizUrl>" title="${layoutSettings.companyName}"><div> </div></a>

Style.css (tomahawk & droppingcrumbs)
Line 945. The logo is used as the background of the logo area, but because the size of the background image means nothing for dimensioning, it will be clipped.
background: url(../images/ofbiz_logo.png) no-repeat;

Solution: This is a hack to put the logo in the style sheet as the background. If the logo is only in the stylesheet without any trick, the logo will be clipped and disappears.
The logo can be loaded as an image directly in header.ftl and it will appear without clip and without any hack in header.ftl.

appbarOpen.ftl (droppingcrumbs)
Line 99, unclosed <li> and <ul>

header.ftl (bizznesstime)
Line 111. div can't be inside span.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@931665 13f79535-47bb-0310-9956-ffa450edef68