As we use the jQuery branch in development of a custom project and we need all to work, I temporarily put back Prototype in the GlobalDecorator. It was needed for Dependent dropdowns mechanism but it's not the only reason, as I will soon transform this mechanism to use jQuery
It was removed at r956423 for https://issues.apache.org/jira/browse/OFBIZ-3817. It can be commented out temporarily while working on stuff where it's easier to not have it in, following Sascha's comment in OFBIZ-3817: <<I removed the prototype loading in the commons screens, the result is that the browser throws an error evertime he find JS code that's not compatible to jQuery. It's easier to find related issues.>>




git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/jquery@992587 13f79535-47bb-0310-9956-ffa450edef68