Feature - Screen Link Ajax Windows (https://issues.apache.org/jira/browse/OFBIZ-4183)
I created a new Screen <link feature. Similar to the Form Lookups a link can be opened in a layered JS Box (see Screenshot in Jira Task). It looks nice and is helpful in some cases.

You can add any screen u want. It's expected that you create a certain screen for the layered link which only contains the information you need.

The layered box doesn't contain any logic it's just a container for the screen which should be opened. So the screen it self can decide what to show. i.e  if you load a form, the form itself implement the way how it react on a submit (either reload the site or perform the submit as Ajax request in the background).

There is also an example in the example section. Try to create a new Example on the mainpage.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1071698 13f79535-47bb-0310-9956-ffa450edef68