Fixed a bug in the getWorkEffortEventsByPeriod service, where the value of the start parameter was being changed - causing incorrect results with periods that do not start at the beginning of the day.
This change might cause problems with client code that assumes the service aligns the start parameter to the start of the day. In those cases, the start parameter alignment should be moved to the client code. See the WorkEffort component Days.groovy, Week.groovy, or Month.groovy scripts for examples.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1223007 13f79535-47bb-0310-9956-ffa450edef68