A patch from Ankit Jain for "User is redirected to showcart page when any error comes on checkout page." https://issues.apache.org/jira/browse/OFBIZ-5206
On checkout page if any error comes it redirects user back to showcart page with an error message.
Expected : Error should be shown on the same page.

To reproduce the issue,
Goto OrderEntry -> Add any product to cart -> click on Quick finalize order now select shipping address & shipping method but don't select any payment option and now click on "Continue to Final Order Review" link. You will be redirected to showcart page and with a error message. "Please select a method of payment."
And expected behaviour is to show the error message on the same checkout page.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1488127 13f79535-47bb-0310-9956-ffa450edef68