Fixes a bug reported by Christian Carlow at https://issues.apache.org/jira/browse/OFBIZ-5457 for "Groovy error when initiating sales order creation for unkown partyId"
This error is generated when an unknown partyId is entered in the Customer field of the sales order entry form:

To resolve the issue, lines 60-63 of CheckInits.groovy should be wrapped in an IF block checking for the existence of the product variable previously assigned.

jleroux: Christian's patch did not apply but it was easy to do it by hand.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1555142 13f79535-47bb-0310-9956-ffa450edef68