Manage multi pk with sub-sequence on entity-auto (OFBIZ-5800),
Add the possibility to the entity-auto engine on the create action to manage entities with more than 2 primary keys which one is under sub sequence or fromDate, like PerfReview (employeePartyId, employeeRoleTypeId, perfReviewId) or PartyQual (partyId, partyQualTypeId, fromDate).

Improve return message for the create action if the entity value exist and the delete action if the entity value not exist instead of the database message error.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1637596 13f79535-47bb-0310-9956-ffa450edef68