A better patch from Adrian for "Field Lookup do not show description when widget-verbose is set to false" https://issues.apache.org/jira/browse/OFBIZ-6396
When widget verbose is set to false and using field lookup, the description of the selected object is not displayed (not using auto completion but the popup).
To reproduce : Set widget.verbose=false in framework/widget/config/widget.properties  
Clear OFBiz cache then go to : https://localhost:8443/partymgr/control/EditPartyRelationships?partyId=Company
In the partyRelationship creation form, in party Id From field, click on lookup, search, select AUTHOR_MADMAX, and so no description displayed. Using autocompletion it's fine.

jleroux: This reverts the previous patch from Gil. It did not occur to me that all the rest of the result was only commented out HTML code, so a simple trim was enough. Happy to get rid of this ugly kludge



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1680881 13f79535-47bb-0310-9956-ffa450edef68