A patch from Florian Montalbano for "Scrum find Total Backlog Item is not working in non-English language" https://issues.apache.org/jira/browse/OFBIZ-7929
The Find form for Product Backlog Item does not work in language other than English when no "statusId" or "custRequestTypeId" are selected.

Step to reproduce :
- Go to the Scrum component and select a product
- Go to the "Total Backlog" tab
or use this link : https://localhost:8443/scrum/control/ViewTotalBacklog?productId=DEMO-PRODUCT-1

- Hit the search button, the find action is performed and return a non-empty list.
- Set your language preference to a non english language (french for example). 
- Click again on the "Total Backlog" tab, and hit the  "Rechercher" button.
- No results are found.

Where does the problem come from ?
The results displayed are from a list named "backlogList" which is built in the following script : FindProductBacklogItem.groovy .
In this script, there were a check on the parameter "custRequestTypeId" and on the parameter "statusId".
But the value was hardcoded :

if("Any".equals(parameters.custRequestTypeId))
and
if(!"Any".equals(parameters.statusId))

Obviously, when the form is in another language than English, those conditions aren't valid anymore (for example, "Any" is "Don't" in French")

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1754051 13f79535-47bb-0310-9956-ffa450edef68