Fixed: "Labels and properties caches clearing no longer work" and "Shortlist the needed resources to start OFBiz in the sourcesets block in Gradle" (OFBIZ-8321)(OFBIZ-7937)
While working on OFBIZ-7798 I noticed an issue with labels cache clearing. 
It still worked correctly in R15.12 and pre-Gradle version 

To test it, load a page with a label, change the label in file, clear all caches
 in webtools, the new label value should appear, it does not.

The same issue exists for properties.

This fixes both and while at it fixes also OFBIZ-7937 "Shortlist the needed 
resources to start OFBiz in the sourcesets block in  Gradle"

I also removed the DTDs from the resources, they seems useless there.

There area also some trailing blanks removed, my Eclipse is set like that.

Thanks: Taher for OFBIZ-7937 description of the issue

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1764280 13f79535-47bb-0310-9956-ffa450edef68