Implemented: Birt Report Builder: an enhancement of the Birt component. Easier user possibility of report creation. (OFBIZ-6919)
The concept, you define a report domain related to an entity or a service through a master content (master report) and a master search form.
When a high level user create a report on a configured domain, OFBiz prepare an instantiation of the master content (flexible report) that permit
to surcharge the search form (recorded dynamically on the content as ElectronicText) and prepare a rptDesign skeleton.
The high level user can download the rptDesign to edit it trhough the birt editor, with all fields that ofbiz will be load as dataset and when is finish upload the result on the content instantiation.
For the end user, he selected the instantiation report, OFBiz display the report search from the content instantiation and the submission call birt with the search parameters and the linked rptDesign.

The rptDesign file is drive by OFBiz to build the data set by two methods :
 * call the performFind if the domain is related to an entity
 * call the a custom service if the domain is related to a service.
For the last case we need two services, a first who prepare and list available fields for search and display. And a second who realize the search.
By convention currently the first service as the same name that the second service suffixed by 'PrepareFields'

This commit contains :
 * useful UI to create a new content instance from a master content. List, edit, remove a content instance and a simple drop-down to list all instance published (ready to use by end user)
 * all services (with the work flexibleReport) to manage a flexible report, prepare the rptDesign skeleton, reanalyze the rptDesign uploaded, prepare the search form
 * Services interface to define you own search custom method
 * Two examples, one on the entity Exemple and one with the service flexibleReportTurnOver. Warning ! there are raw example without a beautiful rptDesign so you need to update the rptDesign with the birt editor to display something.
Thanks: François Wurmser for initialize this improvement, Jacques and Gil for their time to analyze, document and refactor the code


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1780683 13f79535-47bb-0310-9956-ffa450edef68